+++ packages/enhanced/src/lib/container/ModuleFederationPlugin.ts
@@ -11,6 +11,7 @@ import {
   composeKeyWithSeparator,
   type moduleFederationPlugin,
 } from '@module-federation/sdk';
+import { PrefetchPlugin } from '@module-federation/data-prefetch/cli';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
 import type { Compiler, WebpackPluginInstance } from 'webpack';
 import SharePlugin from '../sharing/SharePlugin';
@@ -19,6 +20,8 @@ import ContainerReferencePlugin from './ContainerReferencePlugin';
 import FederationRuntimePlugin from './runtime/FederationRuntimePlugin';
 import { RemoteEntryPlugin } from './runtime/RemoteEntryPlugin';
 import { ExternalsType } from 'webpack/declarations/WebpackOptions';
+import StartupChunkDependenciesPlugin from '../startup/MfStartupChunkDependenciesPlugin';
+import FederationModulesPlugin from './runtime/FederationModulesPlugin';
 
 const isValidExternalsType = require(
   normalizeWebpackPath(
@@ -63,10 +66,20 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
         compiler,
       );
     }
+    if (options.experiments?.federationRuntime) {
+      new FederationModulesPlugin().apply(compiler);
+      new StartupChunkDependenciesPlugin({
+        asyncChunkLoading: true,
+      }).apply(compiler);
+    }
+
     if (options.dts !== false) {
       new DtsPlugin(options).apply(compiler);
     }
+    new PrefetchPlugin(options).apply(compiler);
+
     new FederationRuntimePlugin(options).apply(compiler);
+
     const library = options.library || { type: 'var', name: options.name };
     const remoteType =
       options.remoteType ||
@@ -115,6 +128,7 @@ class ModuleFederationPlugin implements WebpackPluginInstance {
           shareScope: options.shareScope,
           exposes: options.exposes!,
           runtimePlugins: options.runtimePlugins,
+          experiments: options.experiments,
         }).apply(compiler);
       }
       if (