+++ packages/enhanced/src/lib/container/runtime/FederationRuntimePlugin.ts
@@ -1,5 +1,12 @@
-import type { Compiler, WebpackPluginInstance } from 'webpack';
+import type {
+  Compiler,
+  WebpackPluginInstance,
+  Compilation,
+  Chunk,
+} from 'webpack';
 import { normalizeWebpackPath } from '@module-federation/sdk/normalize-webpack-path';
+import { PrefetchPlugin } from '@module-federation/data-prefetch/cli';
+import { moduleFederationPlugin } from '@module-federation/sdk';
 import FederationRuntimeModule from './FederationRuntimeModule';
 import {
   getFederationGlobalScope,
@@ -11,8 +18,15 @@ import {
 import fs from 'fs';
 import path from 'path';
 import { TEMP_DIR } from '../constant';
-import type { moduleFederationPlugin } from '@module-federation/sdk';
+import EmbedFederationRuntimePlugin from './EmbedFederationRuntimePlugin';
+import FederationModulesPlugin from './FederationModulesPlugin';
+import HoistContainerReferences from '../HoistContainerReferencesPlugin';
 import pBtoa from 'btoa';
+import FederationRuntimeDependency from './FederationRuntimeDependency';
+
+const ModuleDependency = require(
+  normalizeWebpackPath('webpack/lib/dependencies/ModuleDependency'),
+) as typeof import('webpack/lib/dependencies/ModuleDependency');
 
 const { RuntimeGlobals, Template } = require(
   normalizeWebpackPath('webpack'),
@@ -32,22 +46,38 @@ const BundlerRuntimePath = require.resolve(
 const RuntimePath = require.resolve('@module-federation/runtime', {
   paths: [RuntimeToolsPath],
 });
+const EmbeddedRuntimePath = require.resolve(
+  '@module-federation/runtime/embedded',
+  {
+    paths: [RuntimeToolsPath],
+  },
+);
 
 const federationGlobal = getFederationGlobalScope(RuntimeGlobals);
+
+const onceForCompler = new WeakSet();
 
 class FederationRuntimePlugin {
   options?: moduleFederationPlugin.ModuleFederationPluginOptions;
   entryFilePath: string;
   bundlerRuntimePath: string;
+  federationRuntimeDependency?: FederationRuntimeDependency; // Add this line
 
   constructor(options?: moduleFederationPlugin.ModuleFederationPluginOptions) {
     this.options = options ? { ...options } : undefined;
     this.entryFilePath = '';
     this.bundlerRuntimePath = BundlerRuntimePath;
+    this.federationRuntimeDependency = undefined; // Initialize as undefined
   }
 
-  static getTemplate(runtimePlugins: string[], bundlerRuntimePath?: string) {
+  static getTemplate(
+    compiler: Compiler,
+    options: moduleFederationPlugin.ModuleFederationPluginOptions,
+    bundlerRuntimePath?: string,
+    experiments?: moduleFederationPlugin.ModuleFederationPluginOptions['experiments'],
+  ) {
     // internal runtime plugin
+    const runtimePlugins = options.runtimePlugins;
     const normalizedBundlerRuntimePath = normalizeToPosixPath(
       bundlerRuntimePath || BundlerRuntimePath,
     );
@@ -68,23 +98,30 @@ class FederationRuntimePlugin {
         runtimePluginNames.push(runtimePluginName);
       });
     }
+    const embedRuntimeLines = Template.asString([
+      `if(!${federationGlobal}.runtime){`,
+      Template.indent([
+        `var prevFederation = ${federationGlobal};`,
+        `${federationGlobal} = {}`,
+        `for(var key in federation){`,
+        Template.indent([`${federationGlobal}[key] = federation[key];`]),
+        '}',
+        `for(var key in prevFederation){`,
+        Template.indent([`${federationGlobal}[key] = prevFederation[key];`]),
+        '}',
+      ]),
+      '}',
+    ]);
 
     return Template.asString([
       `import federation from '${normalizedBundlerRuntimePath}';`,
       runtimePluginTemplates,
-      `var prevFederation = ${federationGlobal};`,
-      `${federationGlobal} = {}`,
-      `for(var key in federation){`,
-      Template.indent([`${federationGlobal}[key] = federation[key];`]),
-      '}',
-      `for(var key in prevFederation){`,
-      Template.indent([`${federationGlobal}[key] = prevFederation[key];`]),
-      '}',
+      embedRuntimeLines,
       `if(!${federationGlobal}.instance){`,
       Template.indent([
         runtimePluginNames.length
           ? Template.asString([
-              `const pluginsToAdd = [`,
+              `var pluginsToAdd = [`,
               Template.indent(
                 runtimePluginNames.map(
                   (item) => `${item} ? (${item}.default || ${item})() : false,`,
@@ -105,25 +142,31 @@ class FederationRuntimePlugin {
         Template.indent([`${federationGlobal}.installInitialConsumes()`]),
         '}',
       ]),
+      PrefetchPlugin.addRuntime(compiler, {
+        name: options.name!,
+      }),
       '}',
     ]);
   }
 
   static getFilePath(
-    containerName: string,
-    runtimePlugins: string[],
+    compiler: Compiler,
+    options: moduleFederationPlugin.ModuleFederationPluginOptions,
     bundlerRuntimePath?: string,
+    experiments?: moduleFederationPlugin.ModuleFederationPluginOptions['experiments'],
   ) {
+    const containerName = options.name;
     const hash = createHash(
       `${containerName} ${FederationRuntimePlugin.getTemplate(
-        runtimePlugins,
+        compiler,
+        options,
         bundlerRuntimePath,
+        experiments,
       )}`,
     );
     return path.join(TEMP_DIR, `entry.${hash}.js`);
   }
-
-  getFilePath() {
+  getFilePath(compiler: Compiler) {
     if (this.entryFilePath) {
       return this.entryFilePath;
     }
@@ -134,26 +177,28 @@ class FederationRuntimePlugin {
 
     if (!this.options?.virtualRuntimeEntry) {
       this.entryFilePath = FederationRuntimePlugin.getFilePath(
-        this.options.name!,
-        this.options.runtimePlugins!,
+        compiler,
+        this.options,
         this.bundlerRuntimePath,
+        this.options.experiments,
       );
     } else {
       this.entryFilePath = `data:text/javascript;charset=utf-8;base64,${pBtoa(
         FederationRuntimePlugin.getTemplate(
-          this.options.runtimePlugins!,
+          compiler,
+          this.options,
           this.bundlerRuntimePath,
+          this.options.experiments,
         ),
       )}`;
     }
     return this.entryFilePath;
   }
-
-  ensureFile() {
+  ensureFile(compiler: Compiler) {
     if (!this.options) {
       return;
     }
-    const filePath = this.getFilePath();
+    const filePath = this.getFilePath(compiler);
     try {
       fs.readFileSync(filePath);
     } catch (err) {
@@ -161,34 +206,84 @@ class FederationRuntimePlugin {
       fs.writeFileSync(
         filePath,
         FederationRuntimePlugin.getTemplate(
-          this.options.runtimePlugins!,
+          compiler,
+          this.options,
           this.bundlerRuntimePath,
+          this.options.experiments,
         ),
       );
     }
   }
 
+  getDependency(compiler: Compiler) {
+    if (this.federationRuntimeDependency)
+      return this.federationRuntimeDependency;
+    this.federationRuntimeDependency = new FederationRuntimeDependency(
+      this.getFilePath(compiler),
+    );
+    return this.federationRuntimeDependency;
+  }
+
   prependEntry(compiler: Compiler) {
     if (!this.options?.virtualRuntimeEntry) {
-      this.ensureFile();
+      this.ensureFile(compiler);
     }
-    const entryFilePath = this.getFilePath();
 
-    modifyEntry({
-      compiler,
-      prependEntry: (entry) => {
-        Object.keys(entry).forEach((entryName) => {
-          const entryItem = entry[entryName];
-          if (!entryItem.import) {
-            // TODO: maybe set this variable as constant is better https://github.com/webpack/webpack/blob/main/lib/config/defaults.js#L176
-            entryItem.import = ['./src'];
-          }
-          if (!entryItem.import.includes(entryFilePath)) {
-            entryItem.import.unshift(entryFilePath);
-          }
-        });
-      },
-    });
+    //if using runtime experiment, use the new include method else patch entry
+    if (this.options?.experiments?.federationRuntime) {
+      compiler.hooks.thisCompilation.tap(
+        this.constructor.name,
+        (compilation: Compilation, { normalModuleFactory }) => {
+          compilation.dependencyFactories.set(
+            FederationRuntimeDependency,
+            normalModuleFactory,
+          );
+          compilation.dependencyTemplates.set(
+            FederationRuntimeDependency,
+            new ModuleDependency.Template(),
+          );
+        },
+      );
+      compiler.hooks.make.tapAsync(
+        this.constructor.name,
+        (compilation: Compilation, callback) => {
+          const federationRuntimeDependency = this.getDependency(compiler);
+          const hooks =
+            FederationModulesPlugin.getCompilationHooks(compilation);
+          compilation.addInclude(
+            compiler.context,
+            federationRuntimeDependency,
+            { name: undefined },
+            (err, module) => {
+              if (err) {
+                return callback(err);
+              }
+              hooks.addFederationRuntimeModule.call(
+                federationRuntimeDependency,
+              );
+              callback();
+            },
+          );
+        },
+      );
+    } else {
+      const entryFilePath = this.getFilePath(compiler);
+      modifyEntry({
+        compiler,
+        prependEntry: (entry) => {
+          Object.keys(entry).forEach((entryName) => {
+            const entryItem = entry[entryName];
+            if (!entryItem.import) {
+              // TODO: maybe set this variable as constant is better https://github.com/webpack/webpack/blob/main/lib/config/defaults.js#L176
+              entryItem.import = ['./src'];
+            }
+            if (!entryItem.import.includes(entryFilePath)) {
+              entryItem.import.unshift(entryFilePath);
+            }
+          });
+        },
+      });
+    }
   }
 
   injectRuntime(compiler: Compiler) {
@@ -205,54 +300,83 @@ class FederationRuntimePlugin {
 
     compiler.hooks.thisCompilation.tap(
       this.constructor.name,
-      (compilation, { normalModuleFactory }) => {
+      (compilation: Compilation) => {
+        const handler = (chunk: Chunk, runtimeRequirements: Set<string>) => {
+          if (runtimeRequirements.has(federationGlobal)) return;
+          runtimeRequirements.add(federationGlobal);
+          runtimeRequirements.add(RuntimeGlobals.interceptModuleExecution);
+          runtimeRequirements.add(RuntimeGlobals.moduleCache);
+          runtimeRequirements.add(RuntimeGlobals.compatGetDefaultExport);
+
+          compilation.addRuntimeModule(
+            chunk,
+            new FederationRuntimeModule(
+              runtimeRequirements,
+              name,
+              initOptionsWithoutShared,
+            ),
+          );
+        };
+
         compilation.hooks.additionalTreeRuntimeRequirements.tap(
           this.constructor.name,
-          (chunk, runtimeRequirements) => {
-            if (runtimeRequirements.has(federationGlobal)) {
+          (chunk: Chunk, runtimeRequirements: Set<string>) => {
+            if (!chunk.hasRuntime()) return;
+            if (runtimeRequirements.has(RuntimeGlobals.initializeSharing))
               return;
-            }
-            runtimeRequirements.add(RuntimeGlobals.interceptModuleExecution);
-            runtimeRequirements.add(RuntimeGlobals.moduleCache);
-            runtimeRequirements.add(RuntimeGlobals.compatGetDefaultExport);
-            runtimeRequirements.add(federationGlobal);
-            compilation.addRuntimeModule(
-              chunk,
-              new FederationRuntimeModule(
-                runtimeRequirements,
-                name,
-                initOptionsWithoutShared,
-              ),
-            );
+            if (runtimeRequirements.has(RuntimeGlobals.currentRemoteGetScope))
+              return;
+            if (runtimeRequirements.has(RuntimeGlobals.shareScopeMap)) return;
+            if (runtimeRequirements.has(federationGlobal)) return;
+            handler(chunk, runtimeRequirements);
           },
         );
+
+        // if federation runtime requirements exist
+        // attach runtime module to the chunk
+        compilation.hooks.runtimeRequirementInTree
+          .for(RuntimeGlobals.initializeSharing)
+          .tap(this.constructor.name, handler);
+        compilation.hooks.runtimeRequirementInTree
+          .for(RuntimeGlobals.currentRemoteGetScope)
+          .tap(this.constructor.name, handler);
+        compilation.hooks.runtimeRequirementInTree
+          .for(RuntimeGlobals.shareScopeMap)
+          .tap(this.constructor.name, handler);
+        compilation.hooks.runtimeRequirementInTree
+          .for(federationGlobal)
+          .tap(this.constructor.name, handler);
       },
     );
   }
 
   setRuntimeAlias(compiler: Compiler) {
-    let runtimePath = RuntimePath;
-    if (this.options?.implementation) {
-      runtimePath = require.resolve('@module-federation/runtime', {
-        paths: [this.options.implementation],
-      });
-    }
-    if (Array.isArray(compiler.options.resolve.alias)) {
-      return;
+    const { experiments, implementation } = this.options || {};
+    const isHoisted = experiments?.federationRuntime === 'hoisted';
+    let runtimePath = isHoisted ? EmbeddedRuntimePath : RuntimePath;
+
+    if (implementation) {
+      runtimePath = require.resolve(
+        `@module-federation/runtime${isHoisted ? '/embedded' : ''}`,
+        { paths: [implementation] },
+      );
     }
 
-    compiler.options.resolve.alias = {
-      ...compiler.options.resolve.alias,
-    };
+    if (isHoisted) {
+      runtimePath = runtimePath.replace('.cjs', '.esm');
+    }
 
-    if (!compiler.options.resolve.alias['@module-federation/runtime$']) {
-      compiler.options.resolve.alias['@module-federation/runtime$'] =
-        runtimePath;
-    }
-    if (!compiler.options.resolve.alias['@module-federation/runtime-tools$']) {
-      compiler.options.resolve.alias['@module-federation/runtime-tools$'] =
-        this.options?.implementation || RuntimeToolsPath;
-    }
+    const alias = compiler.options.resolve.alias || {};
+    alias['@module-federation/runtime$'] =
+      alias['@module-federation/runtime$'] || runtimePath;
+    alias['@module-federation/runtime-tools$'] =
+      alias['@module-federation/runtime-tools$'] ||
+      implementation ||
+      RuntimeToolsPath;
+
+    // Set up aliases for the federation runtime and tools
+    // This ensures that the correct versions are used throughout the project
+    compiler.options.resolve.alias = alias;
   }
 
   apply(compiler: Compiler) {
@@ -281,7 +405,6 @@ class FederationRuntimePlugin {
     );
 
     if (useContainerPlugin && !this.options) {
-      // @ts-ignore
       this.options = useContainerPlugin._options;
     }
 
@@ -292,7 +415,7 @@ class FederationRuntimePlugin {
       };
     }
     if (this.options && !this.options?.name) {
-      // the instance may get the same one if the name is the same https://github.com/module-federation/core/blob/main/packages/runtime/src/index.ts#L18
+      //! the instance may get the same one if the name is the same https://github.com/module-federation/core/blob/main/packages/runtime/src/index.ts#L18
       this.options.name =
         compiler.options.output.uniqueName || `container_${Date.now()}`;
     }
@@ -306,9 +429,36 @@ class FederationRuntimePlugin {
       );
     }
 
-    this.prependEntry(compiler);
-    this.injectRuntime(compiler);
-    this.setRuntimeAlias(compiler);
+    if (this.options?.experiments?.federationRuntime === 'hoisted') {
+      this.bundlerRuntimePath = this.bundlerRuntimePath.replace(
+        '.cjs.js',
+        '.esm.js',
+      );
+
+      new EmbedFederationRuntimePlugin().apply(compiler);
+
+      new HoistContainerReferences().apply(compiler);
+
+      new compiler.webpack.NormalModuleReplacementPlugin(
+        /@module-federation\/runtime/,
+        (resolveData) => {
+          if (/webpack-bundler-runtime/.test(resolveData.contextInfo.issuer)) {
+            resolveData.request = RuntimePath.replace('cjs', 'esm');
+
+            if (resolveData.createData) {
+              resolveData.createData.request = resolveData.request;
+            }
+          }
+        },
+      ).apply(compiler);
+    }
+    // dont run multiple times on every apply()
+    if (!onceForCompler.has(compiler)) {
+      this.prependEntry(compiler);
+      this.injectRuntime(compiler);
+      this.setRuntimeAlias(compiler);
+      onceForCompler.add(compiler);
+    }
   }
 }