+++ packages/runtime-tools/src/runtime.ts
@@ -1,2 +1,1 @@
 export * from '@module-federation/runtime';
-export { default as RetryPlugin } from '@module-federation/runtime/retry-plugin';